Refactor verification to run scripts dynamically and expose simple api
This is achieved by:
- Adding seperate scripts for validating recipes and packages
- Using verify class to run verification scripts
- Deleting unused modules (package and recipe) in favour of new architecture
- renaming anaconda_verify to conda_verify
