+++ .gitignore
@@ -3,3 +3,4 @@
 *.egg-info
 .cache/
 build/
+.idea/
