+++ conda_verify/exceptions.py
@@ -1,0 +1,6 @@
+class RecipeError(Exception):
+    pass
+
+
+class PackageError(Exception):
+    pass
