+++ conda_verify/package/test_no_2to3_pickle.py
@@ -1,0 +1,7 @@
+from conda_verify.conda_package_check import CondaPackageCheck
+
+
+def verify(path_to_package=None, verbose=True, **kwargs):
+    package_check = CondaPackageCheck(path_to_package, verbose)
+    package_check.no_2to3_pickle()
+    package_check.t.close()
