+++ setup.py
@@ -2,8 +2,12 @@
 # All Rights Reserved
 import re
 from os.path import join
+import sys
 
-from distutils.core import setup
+if 'develop' in sys.argv:
+    from setuptools import setup
+else:
+    from distutils.core import setup
 
 
 # read version from conda_verify/__init__.py
